+++ internal/exec/exec.go
@@ -107,6 +107,20 @@ func RunnerWithPrintFields(printFields string) RunnerOption {
 	}
 }
 
+// RunnerWithProtocBinPath returns a RunnerOption that uses the given protoc binary path.
+func RunnerWithProtocBinPath(protocBinPath string) RunnerOption {
+	return func(runner *runner) {
+		runner.protocBinPath = protocBinPath
+	}
+}
+
+// RunnerWithProtocWKTPath returns a RunnerOption that uses the given path to include the well-known types.
+func RunnerWithProtocWKTPath(protocWKTPath string) RunnerOption {
+	return func(runner *runner) {
+		runner.protocWKTPath = protocWKTPath
+	}
+}
+
 // RunnerWithProtocURL returns a RunnerOption that uses the given protoc zip file URL.
 func RunnerWithProtocURL(protocURL string) RunnerOption {
 	return func(runner *runner) {