+++ internal/protoc/compiler.go
@@ -68,6 +68,8 @@ var (
 type compiler struct {
 	logger              *zap.Logger
 	cachePath           string
+	protocBinPath       string
+	protocWKTPath       string
 	protocURL           string
 	doGen               bool
 	doFileDescriptorSet bool
@@ -268,7 +270,10 @@ func (c *compiler) getCmdMetas(protoSet *file.ProtoSet) (cmdMetas []*cmdMeta, re
 	}()
 	// you need a new downloader for every ProtoSet as each configuration file could
 	// have a different protoc.version value
-	downloader := c.newDownloader(protoSet.Config)
+	downloader, err := c.newDownloader(protoSet.Config)
+	if err != nil {
+		return nil, err
+	}
 	if _, err := downloader.Download(); err != nil {
 		return cmdMetas, err
 	}
@@ -354,7 +359,7 @@ func (c *compiler) getCmdMetas(protoSet *file.ProtoSet) (cmdMetas []*cmdMeta, re
 	return cmdMetas, nil
 }
 
-func (c *compiler) newDownloader(config settings.Config) Downloader {
+func (c *compiler) newDownloader(config settings.Config) (Downloader, error) {
 	downloaderOptions := []DownloaderOption{
 		DownloaderWithLogger(c.logger),
 	}
@@ -362,6 +367,18 @@ func (c *compiler) newDownloader(config settings.Config) Downloader {
 		downloaderOptions = append(
 			downloaderOptions,
 			DownloaderWithCachePath(c.cachePath),
+		)
+	}
+	if c.protocBinPath != "" {
+		downloaderOptions = append(
+			downloaderOptions,
+			DownloaderWithProtocBinPath(c.protocBinPath),
+		)
+	}
+	if c.protocWKTPath != "" {
+		downloaderOptions = append(
+			downloaderOptions,
+			DownloaderWithProtocWKTPath(c.protocWKTPath),
 		)
 	}
 	if c.protocURL != "" {