+++ internal/protoc/downloader.go
@@ -43,17 +43,24 @@ import (
 )
 
 type downloader struct {
+	lock sync.RWMutex
+
 	logger    *zap.Logger
 	cachePath string
 	protocURL string
 	config    settings.Config
 
-	lock sync.RWMutex
 	// the looked-up and verified to exist base path
 	cachedBasePath string
+
+	// If set, Prototool will invoke protoc and include
+	// the well-known-types, from the configured binPath
+	// and wktPath.
+	protocBinPath string
+	protocWKTPath string
 }
 
-func newDownloader(config settings.Config, options ...DownloaderOption) *downloader {
+func newDownloader(config settings.Config, options ...DownloaderOption) (*downloader, error) {
 	downloader := &downloader{
 		config: config,
 		logger: zap.NewNop(),
@@ -64,7 +71,33 @@ func newDownloader(config settings.Config, options ...DownloaderOption) *downloa
 	if downloader.config.Compile.ProtobufVersion == "" {
 		downloader.config.Compile.ProtobufVersion = vars.DefaultProtocVersion
 	}
-	return downloader
+	if downloader.protocBinPath != "" || downloader.protocWKTPath != "" {
+		if downloader.protocURL != "" {
+			return nil, fmt.Errorf("cannot use protoc-url in combination with either protoc-bin-path or protoc-wkt-path")
+		}
+		if downloader.protocBinPath == "" || downloader.protocWKTPath == "" {
+			return nil, fmt.Errorf("both protoc-bin-path and protoc-wkt-path must be set")
+		}
+		cleanBinPath := filepath.Clean(downloader.protocBinPath)
+		if _, err := os.Stat(cleanBinPath); os.IsNotExist(err) {
+			return nil, err
+		}
+		cleanWKTPath := filepath.Clean(downloader.protocWKTPath)
+		if _, err := os.Stat(cleanWKTPath); os.IsNotExist(err) {
+			return nil, err
+		}
+		protobufPath := filepath.Join(cleanWKTPath, "google", "protobuf")
+		info, err := os.Stat(protobufPath)
+		if os.IsNotExist(err) {
+			return nil, err
+		}
+		if !info.IsDir() {
+			return nil, fmt.Errorf("%q is not a valid well-known types directory", protobufPath)
+		}
+		downloader.protocBinPath = cleanBinPath
+		downloader.protocWKTPath = cleanWKTPath
+	}
+	return downloader, nil
 }
 
 func (d *downloader) Download() (string, error) {
@@ -78,6 +111,9 @@ func (d *downloader) Download() (string, error) {
 }
 
 func (d *downloader) ProtocPath() (string, error) {
+	if d.protocBinPath != "" {
+		return d.protocBinPath, nil
+	}
 	basePath, err := d.Download()
 	if err != nil {
 		return "", err
@@ -86,6 +122,9 @@ func (d *downloader) ProtocPath() (string, error) {
 }
 
 func (d *downloader) WellKnownTypesIncludePath() (string, error) {
+	if d.protocWKTPath != "" {
+		return d.protocWKTPath, nil
+	}
 	basePath, err := d.Download()
 	if err != nil {
 		return "", err
@@ -104,6 +143,10 @@ func (d *downloader) Delete() error {
 }
 
 func (d *downloader) cache() (string, error) {
+	if d.protocBinPath != "" {
+		return d.protocBinPath, nil
+	}
+
 	d.lock.Lock()
 	defer d.lock.Unlock()