+++ internal/protoc/protoc.go
@@ -84,6 +84,21 @@ func DownloaderWithCachePath(cachePath string) DownloaderOption {
 	}
 }
 
+// DownloaderWithProtocBinPath returns a DownloaderOption that uses the given protoc binary path.
+func DownloaderWithProtocBinPath(protocBinPath string) DownloaderOption {
+	return func(downloader *downloader) {
+		downloader.protocBinPath = protocBinPath
+	}
+}
+
+// DownloaderWithProtocWKTPath returns a DownloaderOption that uses the given path to include
+// the well-known types.
+func DownloaderWithProtocWKTPath(protocWKTPath string) DownloaderOption {
+	return func(downloader *downloader) {
+		downloader.protocWKTPath = protocWKTPath
+	}
+}
+
 // DownloaderWithProtocURL returns a DownloaderOption that uses the given protoc zip file URL.
 //
 // The default is https://github.com/protocolbuffers/protobuf/releases/download/vVERSION/protoc-VERSION-OS-ARCH.zip.
@@ -94,7 +109,7 @@ func DownloaderWithProtocURL(protocURL string) DownloaderOption {
 }
 
 // NewDownloader returns a new Downloader for the given config and DownloaderOptions.
-func NewDownloader(config settings.Config, options ...DownloaderOption) Downloader {
+func NewDownloader(config settings.Config, options ...DownloaderOption) (Downloader, error) {
 	return newDownloader(config, options...)
 }
 
@@ -143,6 +158,22 @@ func CompilerWithLogger(logger *zap.Logger) CompilerOption {
 func CompilerWithCachePath(cachePath string) CompilerOption {
 	return func(compiler *compiler) {
 		compiler.cachePath = cachePath
+	}
+}
+
+// CompilerWithProtocBinPath returns a CompilerOption that uses the given protoc binary path.
+//
+func CompilerWithProtocBinPath(protocBinPath string) CompilerOption {
+	return func(compiler *compiler) {
+		compiler.protocBinPath = protocBinPath
+	}
+}
+
+// CompilerWithProtocWKTPath returns a CompilerOption that uses the given path to include the
+// well-known types.
+func CompilerWithProtocWKTPath(protocWKTPath string) CompilerOption {
+	return func(compiler *compiler) {
+		compiler.protocWKTPath = protocWKTPath
 	}
 }